Allow traffic load to be disabled
We want to be able to run this process on staging and
integration without doing the data download from GA.

Passing in the `SKIP_TRAFFIC_LOAD` will allow us to do this

+++ nightly-run.sh
@@ -2,13 +2,15 @@
 set -e
 set -o pipefail
 
-if [ \! -d ENV ]; then virtualenv ENV; fi
-. ENV/bin/activate
-pip install -r requirements.txt
-rm -f page-traffic.dump
-PYTHONPATH=. python scripts/fetch.py page-traffic.dump 14
-SEARCH_NODE=$(/usr/local/bin/govuk_node_list -c search --single-node)
-ssh deploy@${SEARCH_NODE} '(cd /var/apps/rummager; govuk_setenv rummager bundle exec ./bin/page_traffic_load)' < page-traffic.dump
+if [ -z $SKIP_TRAFFIC_LOAD ]; then
+  if [ \! -d ENV ]; then virtualenv ENV; fi
+  . ENV/bin/activate
+  pip install -r requirements.txt
+  rm -f page-traffic.dump
+  PYTHONPATH=. python scripts/fetch.py page-traffic.dump 14
+  SEARCH_NODE=$(/usr/local/bin/govuk_node_list -c search --single-node)
+  ssh deploy@${SEARCH_NODE} '(cd /var/apps/rummager; govuk_setenv rummager bundle exec ./bin/page_traffic_load)' < page-traffic.dump
+fi
 
 ssh deploy@${SEARCH_NODE} '(cd /var/apps/rummager; PROCESS_ALL_DATA=true RUMMAGER_INDEX=mainstream govuk_setenv rummager bundle exec rake rummager:update_popularity)'
 ssh deploy@${SEARCH_NODE} '(cd /var/apps/rummager; PROCESS_ALL_DATA=true RUMMAGER_INDEX=detailed govuk_setenv rummager bundle exec rake rummager:update_popularity)'
